add category management in menu editor
add an optional value for the dine in option, thats allows to select the number of guest

change de drawer menu to keep just one logo text instead of 2

add a staff management role type of kitchen

in kitchen display module, as the bump order button add the smark ready or start prep or ready button. show the current status as is, but the button that can change it make it different. also add a button that allows to bump all orders. also link the kitchen display order from de active orders list, and add a kitchen selector if there is more than one kitchen

on settings module on general section add something to activate or deactivate the kitchen display